hal_nxp: Sync MCXW23 header files in 25.09-pvw2 to hal_nxp
1. Removed library and made power and system sources public
2. Enabled the internal registers for power driver

+++ mcux/mcux-sdk-ng/devices/MCX/MCXW/MCXW236/drivers/power_manager_mcxw23/device/fsl_pm_device.c
@@ -450,8 +450,8 @@ void PMDEVICE_StopSleepTimer(void)
 uint64_t PMDEVICE_GetSleepTimer(void)
 {
 #ifdef TIMER_PORT_TYPE_CTIMER
-    return (CTIMER_GetTimerCountValue(CTIMER0) * 1e6 / CLOCK_GetCTimerClkFreq(0));
+    return (uint64_t)(CTIMER_GetTimerCountValue(CTIMER0) * 1e6 / CLOCK_GetCTimerClkFreq(0));
 #else
-    return (OSTIMER_GetCurrentTimerValue(OSTIMER) * 1e6 / CLOCK_GetOSTimerClkFreq());
+    return (uint64_t)(OSTIMER_GetCurrentTimerValue(OSTIMER) * 1e6 / CLOCK_GetOSTimerClkFreq());
 #endif /* TIMER_PORT_TYPE_CTIMER */
 }
